IDVA6-1181 Test files moved to the correct directory
